CentralServer/pyproject.toml
[project]
name = "centralserver"
version = "0.5.0"
description = "The central server for the project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "charset-normalizer>=3.4.2",
    "concurrent-log-handler>=0.9.28",
    "dateparser>=1.2.2",
    "fastapi[standard]>=0.115.12",
    "google-generativeai>=0.8.5",
    "jinja2>=3.1.6",
    "llm>=0.26",
    "minio>=7.2.15",
    "passlib[argon2]>=1.7.4",
    "pillow>=11.2.1",
    "psycopg[binary]>=3.2.9",
    "pymysql>=1.1.1",
    "pyotp>=2.9.0",
    "pyspellchecker>=0.8.3",
    "python-dateutil>=2.9.0.post0",
    "python-jose[cryptography]>=3.4.0",
    "pytz>=2025.2",
    "rapidfuzz>=3.13.0",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.40",
    "sqlmodel>=0.0.24",
    "types-pillow>=10.2.0.20240822",
    "uvicorn>=0.34.0",
]

[project.optional-dependencies]
ai = ["llm-gemini>=0.1.0"]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "faker>=37.4.0",
    "httpx>=0.28.1",
    "isort>=6.0.1",
    "pylint>=3.3.6",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.0",
    "pytest-dotenv>=0.5.2",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--cache-clear --color=yes --code-highlight=yes --cov=centralserver --cov-branch --cov-report=term --cov-report=xml -vvv"
testpaths = ["tests"]
env_files = [".pytest.env"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
